Population test data with full count stratum

Test population data for using in struktuR package (synthetic data). Stratum "B" is a full count strata.

Usage

pop_data_fulltelling

Format

A data frame with 8065 rows and 7 variables:

id

Identification variable

employees

Number of employees

employees_f

number of female employees

employees_m

number of male employees

turnover

Amount of turnover of the company (in 000 NOK)

size

Category size of the company in terms of turnover (small, mid, large)

industry

The activity group for which the business works in.